Split widget: filter chip
