+++ extension/devtools/app/lib/main.dart
@@ -20,6 +20,7 @@ part 'hooks/use_timeline_panel_state.dart';
 part 'hooks/use_ui_state.dart';
 part 'utils/helpers.dart';
 part 'widgets/action_pill.dart';
+part 'widgets/filter_chip.dart';
 part 'widgets/status_badge.dart';
 
 void main() {
@@ -3177,48 +3178,6 @@ class _SignalDetail extends StatelessWidget {
             Text(entry!.note, style: textTheme.bodySmall),
           ],
         ],
-      ),
-    );
-  }
-}
-
-class _FilterChip extends StatelessWidget {
-  const _FilterChip({
-    required this.label,
-    required this.isSelected,
-    required this.onTap,
-  });
-
-  final String label;
-  final bool isSelected;
-  final VoidCallback onTap;
-
-  @override
-  Widget build(BuildContext context) {
-    final colorScheme = Theme.of(context).colorScheme;
-    final background = isSelected
-        ? OrefPalette.indigo.withValues(alpha: 0.35)
-        : colorScheme.surface.withValues(alpha: 0.35);
-
-    return Material(
-      color: Colors.transparent,
-      child: InkWell(
-        borderRadius: BorderRadius.circular(999),
-        onTap: onTap,
-        child: AnimatedContainer(
-          duration: const Duration(milliseconds: 150),
-          padding: const EdgeInsets.symmetric(horizontal: 14, vertical: 8),
-          decoration: BoxDecoration(
-            color: background,
-            borderRadius: BorderRadius.circular(999),
-            border: Border.all(
-              color: isSelected
-                  ? OrefPalette.indigo.withValues(alpha: 0.7)
-                  : colorScheme.onSurface.withValues(alpha: 0.1),
-            ),
-          ),
-          child: Text(label, style: Theme.of(context).textTheme.labelMedium),
-        ),
       ),
     );
   }
