+++ extension/devtools/app/lib/widgets/filter_chip.dart
@@ -1,0 +1,43 @@
+part of '../main.dart';
+
+class _FilterChip extends StatelessWidget {
+  const _FilterChip({
+    required this.label,
+    required this.isSelected,
+    required this.onTap,
+  });
+
+  final String label;
+  final bool isSelected;
+  final VoidCallback onTap;
+
+  @override
+  Widget build(BuildContext context) {
+    final colorScheme = Theme.of(context).colorScheme;
+    final background = isSelected
+        ? OrefPalette.indigo.withValues(alpha: 0.35)
+        : colorScheme.surface.withValues(alpha: 0.35);
+
+    return Material(
+      color: Colors.transparent,
+      child: InkWell(
+        borderRadius: BorderRadius.circular(999),
+        onTap: onTap,
+        child: AnimatedContainer(
+          duration: const Duration(milliseconds: 150),
+          padding: const EdgeInsets.symmetric(horizontal: 14, vertical: 8),
+          decoration: BoxDecoration(
+            color: background,
+            borderRadius: BorderRadius.circular(999),
+            border: Border.all(
+              color: isSelected
+                  ? OrefPalette.indigo.withValues(alpha: 0.7)
+                  : colorScheme.onSurface.withValues(alpha: 0.1),
+            ),
+          ),
+          child: Text(label, style: Theme.of(context).textTheme.labelMedium),
+        ),
+      ),
+    );
+  }
+}
